refactor: typify utils (#171)
* Typify custom errors

* Typify permission levels

* Move errors to their own files

* Update file paths

* Update WebhookValidationError to be extended from ApplicationError

* Update ConflictError to extend from ApplicationError

* JSDoc the custom errors

* Fix rebase errors
