--- src/app/utils/permission-levels.js
@@ -1,7 +1,0 @@
-const PERMISSION_LEVELS = {
-  READ: 'read',
-  WRITE: 'write',
-  DELETE: 'delete',
-}
-
-module.exports = PERMISSION_LEVELS
