+++ src/app/utils/permission-levels.ts
@@ -1,0 +1,7 @@
+enum PERMISSION_LEVELS {
+  READ = 'read',
+  WRITE = 'write',
+  DELETE = 'delete',
+}
+
+export default PERMISSION_LEVELS
